tests/resources/collection-generator/pyproject.toml
[project]
name = "collection-generator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "mutagen>=1.47.0",
    "piper-tts>=1.3.0",
]
